weather status and animation added

+++ app/src/main/java/com/example/wheather_app/WeatherModal/Current.java
@@ -40,8 +40,8 @@ public class Current {
     public void setTempC(int tempC) {
         this.tempC = tempC;
     }
-    public float getTempC() {
-        return tempC;
+    public int getTempC() {
+        return (int)tempC;
     }
